Unified alerts' API and fetcher specific type methods + split DataMonitoringAlerts flow into steps (populate, fetch, filter, format, skip, send)
