+++ elementary/monitor/fetchers/alerts/alerts.py
@@ -1,5 +1,5 @@
 import json
-from typing import Dict, List, Optional, Union
+from typing import Dict, List, Optional
 
 from elementary.clients.dbt.base_dbt_runner import BaseDbtRunner
 from elementary.clients.fetcher.fetcher import FetcherClient
@@ -7,9 +7,6 @@ from elementary.config.config import Config
 from elementary.monitor.fetchers.alerts.schema.pending_alerts import (
     AlertTypes,
     PendingAlertSchema,
-    PendingModelAlertSchema,
-    PendingSourceFreshnessAlertSchema,
-    PendingTestAlertSchema,
 )
 from elementary.utils.log import get_logger
 from elementary.utils.time import get_now_utc_str
@@ -30,11 +27,7 @@ class AlertsFetcher(FetcherClient):
 
     def skip_alerts(
         self,
-        alerts_to_skip: Union[
-            List[PendingTestAlertSchema],
-            List[PendingModelAlertSchema],
-            List[PendingSourceFreshnessAlertSchema],
-        ],
+        alerts_to_skip: List[PendingAlertSchema],
     ):
         alert_ids = [alert.id for alert in alerts_to_skip]
         alert_ids_chunks = self._split_list_to_chunks(alert_ids)
@@ -46,7 +39,7 @@ class AlertsFetcher(FetcherClient):
                 quiet=True,
             )
 
-    def _query_pending_alerts(
+    def query_pending_alerts(
         self, days_back: int, type: Optional[AlertTypes] = None
     ) -> List[PendingAlertSchema]:
         pending_alerts_results = self.dbt_runner.run_operation(
@@ -58,7 +51,7 @@ class AlertsFetcher(FetcherClient):
             for result in json.loads(pending_alerts_results[0])
         ]
 
-    def _query_last_alert_times(
+    def query_last_alert_times(
         self, days_back: int, type: Optional[AlertTypes] = None
     ) -> Dict[str, str]:
         response = self.dbt_runner.run_operation(
@@ -66,54 +59,6 @@ class AlertsFetcher(FetcherClient):
             macro_args={"days_back": days_back, "type": type.value if type else None},
         )
         return json.loads(response[0])
-
-    def query_pending_test_alerts(self, days_back: int) -> List[PendingTestAlertSchema]:
-        logger.info("Querying test alerts.")
-        pending_test_alerts_results = self._query_pending_alerts(
-            days_back=days_back, type=AlertTypes.TEST
-        )
-        return [
-            PendingTestAlertSchema(**result.data)
-            for result in pending_test_alerts_results
-        ]
-
-    def query_pending_model_alerts(
-        self, days_back: int
-    ) -> List[PendingModelAlertSchema]:
-        logger.info("Querying model alerts.")
-        pending_model_alerts_results = self._query_pending_alerts(
-            days_back=days_back, type=AlertTypes.MODEL
-        )
-        return [
-            PendingModelAlertSchema(**result.data)
-            for result in pending_model_alerts_results
-        ]
-
-    def query_pending_source_freshness_alerts(
-        self, days_back: int
-    ) -> List[PendingSourceFreshnessAlertSchema]:
-        logger.info("Querying source freshness alerts.")
-        pending_source_freshness_alerts_results = self._query_pending_alerts(
-            days_back=days_back, type=AlertTypes.SOURCE_FRESHNESS
-        )
-        return [
-            PendingSourceFreshnessAlertSchema(**result.data)
-            for result in pending_source_freshness_alerts_results
-        ]
-
-    def query_last_test_alert_times(self, days_back: int) -> Dict[str, str]:
-        logger.info("Querying test alerts last sent times.")
-        return self._query_last_alert_times(days_back=days_back, type=AlertTypes.TEST)
-
-    def query_last_model_alert_times(self, days_back: int) -> Dict[str, str]:
-        logger.info("Querying model alerts last sent times.")
-        return self._query_last_alert_times(days_back=days_back, type=AlertTypes.MODEL)
-
-    def query_last_source_freshness_alert_times(self, days_back: int) -> Dict[str, str]:
-        logger.info("Querying source freshness alerts last sent times.")
-        return self._query_last_alert_times(
-            days_back=days_back, type=AlertTypes.SOURCE_FRESHNESS
-        )
 
     def update_sent_alerts(self, alert_ids: List[str]) -> None:
         alert_ids_chunks = self._split_list_to_chunks(alert_ids)
