+++ tests/unit/monitor/api/alerts/test_alert_filters.py
@@ -18,257 +18,346 @@ from elementary.monitor.data_monitoring.schema import (
     SupportedFilterTypes,
 )
 from elementary.monitor.fetchers.alerts.schema.pending_alerts import (
-    PendingModelAlertSchema,
-    PendingSourceFreshnessAlertSchema,
-    PendingTestAlertSchema,
+    AlertTypes,
+    PendingAlertSchema,
 )
 
 
 def initial_alerts():
     test_alerts = [
-        PendingTestAlertSchema(
+        PendingAlertSchema(
             id="1",
             alert_class_id="test_id_1",
-            model_unique_id="elementary.model_id_1",
-            test_unique_id="test_id_1",
-            test_name="test_1",
-            test_created_at="2022-10-10 10:10:10",
-            tags='["one", "two"]',
-            model_meta=dict(owner='["jeff", "john"]'),
-            status="fail",
-            elementary_unique_id="elementary.model_id_1.test_id_1.9cf2f5f6ad.None.generic",
+            type=AlertTypes.TEST,
             detected_at="2022-10-10 10:00:00",
-            database_name="test_db",
-            schema_name="test_schema",
-            table_name="table",
-            suppression_status="pending",
-            test_type="dbt_test",
-            test_sub_type="generic",
-            test_results_description="a mock alert",
-            test_results_query="select * from table",
-            test_short_name="test_1",
-            severity="ERROR",
+            created_at="2022-10-10 10:00:00",
+            updated_at="2022-10-10 10:00:00",
+            status="pending",
+            data=dict(
+                id="1",
+                alert_class_id="test_id_1",
+                model_unique_id="elementary.model_id_1",
+                test_unique_id="test_id_1",
+                test_name="test_1",
+                test_created_at="2022-10-10 10:10:10",
+                tags='["one", "two"]',
+                model_meta=dict(owner='["jeff", "john"]'),
+                status="fail",
+                elementary_unique_id="elementary.model_id_1.test_id_1.9cf2f5f6ad.None.generic",
+                detected_at="2022-10-10 10:00:00",
+                database_name="test_db",
+                schema_name="test_schema",
+                table_name="table",
+                suppression_status="pending",
+                test_type="dbt_test",
+                test_sub_type="generic",
+                test_results_description="a mock alert",
+                test_results_query="select * from table",
+                test_short_name="test_1",
+                severity="ERROR",
+            ),
         ),
-        PendingTestAlertSchema(
+        PendingAlertSchema(
             id="2",
             alert_class_id="test_id_2",
-            model_unique_id="elementary.model_id_1",
-            test_unique_id="test_id_2",
-            test_name="test_2",
-            test_created_at="2022-10-10 09:10:10",
-            tags='["three"]',
-            model_meta=dict(owner='["jeff", "john"]'),
-            status="fail",
-            elementary_unique_id="elementary.model_id_1.test_id_2.9cf2f5f6ad.None.generic",
+            type=AlertTypes.TEST,
             detected_at="2022-10-10 10:00:00",
-            database_name="test_db",
-            schema_name="test_schema",
-            table_name="table",
-            suppression_status="pending",
-            test_type="dbt_test",
-            test_sub_type="generic",
-            test_results_description="a mock alert",
-            test_results_query="select * from table",
-            test_short_name="test_2",
-            severity="ERROR",
+            created_at="2022-10-10 10:00:00",
+            updated_at="2022-10-10 10:00:00",
+            status="pending",
+            data=dict(
+                id="2",
+                alert_class_id="test_id_2",
+                model_unique_id="elementary.model_id_1",
+                test_unique_id="test_id_2",
+                test_name="test_2",
+                test_created_at="2022-10-10 09:10:10",
+                tags='["three"]',
+                model_meta=dict(owner='["jeff", "john"]'),
+                status="fail",
+                elementary_unique_id="elementary.model_id_1.test_id_2.9cf2f5f6ad.None.generic",
+                detected_at="2022-10-10 10:00:00",
+                database_name="test_db",
+                schema_name="test_schema",
+                table_name="table",
+                suppression_status="pending",
+                test_type="dbt_test",
+                test_sub_type="generic",
+                test_results_description="a mock alert",
+                test_results_query="select * from table",
+                test_short_name="test_2",
+                severity="ERROR",
+            ),
         ),
-        PendingTestAlertSchema(
+        PendingAlertSchema(
             id="3",
             alert_class_id="test_id_3",
-            model_unique_id="elementary.model_id_2",
-            test_unique_id="test_id_3",
-            test_name="test_3",
-            test_created_at="2022-10-10 10:10:10",
-            # invalid tag
-            tags="one",
-            model_meta=dict(owner='["john"]'),
-            status="fail",
-            elementary_unique_id="elementary.model_id_1.test_id_3.9cf2f5f6ad.None.generic",
+            type=AlertTypes.TEST,
             detected_at="2022-10-10 10:00:00",
-            database_name="test_db",
-            schema_name="test_schema",
-            table_name="table",
-            suppression_status="pending",
-            test_type="dbt_test",
-            test_sub_type="generic",
-            test_results_description="a mock alert",
-            test_results_query="select * from table",
-            test_short_name="test_3",
-            severity="ERROR",
+            created_at="2022-10-10 10:00:00",
+            updated_at="2022-10-10 10:00:00",
+            status="pending",
+            data=dict(
+                id="3",
+                alert_class_id="test_id_3",
+                model_unique_id="elementary.model_id_2",
+                test_unique_id="test_id_3",
+                test_name="test_3",
+                test_created_at="2022-10-10 10:10:10",
+                # invalid tag
+                tags="one",
+                model_meta=dict(owner='["john"]'),
+                status="fail",
+                elementary_unique_id="elementary.model_id_1.test_id_3.9cf2f5f6ad.None.generic",
+                detected_at="2022-10-10 10:00:00",
+                database_name="test_db",
+                schema_name="test_schema",
+                table_name="table",
+                suppression_status="pending",
+                test_type="dbt_test",
+                test_sub_type="generic",
+                test_results_description="a mock alert",
+                test_results_query="select * from table",
+                test_short_name="test_3",
+                severity="ERROR",
+            ),
         ),
-        PendingTestAlertSchema(
+        PendingAlertSchema(
             id="4",
             alert_class_id="test_id_4",
-            model_unique_id="elementary.model_id_2",
-            test_unique_id="test_id_4",
-            test_name="test_4",
-            test_created_at="2022-10-10 09:10:10",
-            tags='["three", "four"]',
-            model_meta=dict(owner='["jeff"]'),
-            status="warn",
-            elementary_unique_id="elementary.model_id_1.test_id_4.9cf2f5f6ad.None.generic",
+            type=AlertTypes.TEST,
             detected_at="2022-10-10 10:00:00",
-            database_name="test_db",
-            schema_name="test_schema",
-            table_name="table",
-            suppression_status="pending",
-            test_type="dbt_test",
-            test_sub_type="generic",
-            test_results_description="a mock alert",
-            test_results_query="select * from table",
-            test_short_name="test_4",
-            severity="ERROR",
+            created_at="2022-10-10 10:00:00",
+            updated_at="2022-10-10 10:00:00",
+            status="pending",
+            data=dict(
+                id="4",
+                alert_class_id="test_id_4",
+                model_unique_id="elementary.model_id_2",
+                test_unique_id="test_id_4",
+                test_name="test_4",
+                test_created_at="2022-10-10 09:10:10",
+                tags='["three", "four"]',
+                model_meta=dict(owner='["jeff"]'),
+                status="warn",
+                elementary_unique_id="elementary.model_id_1.test_id_4.9cf2f5f6ad.None.generic",
+                detected_at="2022-10-10 10:00:00",
+                database_name="test_db",
+                schema_name="test_schema",
+                table_name="table",
+                suppression_status="pending",
+                test_type="dbt_test",
+                test_sub_type="generic",
+                test_results_description="a mock alert",
+                test_results_query="select * from table",
+                test_short_name="test_4",
+                severity="ERROR",
+            ),
         ),
     ]
     model_alerts = [
-        PendingModelAlertSchema(
+        PendingAlertSchema(
             id="1",
             alert_class_id="elementary.model_id_1",
-            model_unique_id="elementary.model_id_1",
-            alias="modely",
-            path="my/path",
-            original_path="",
-            materialization="table",
-            message="",
-            full_refresh=False,
+            type=AlertTypes.MODEL,
             detected_at="2022-10-10 10:00:00",
-            alert_suppression_interval=0,
-            tags='["one", "two"]',
-            model_meta=dict(owner='["jeff", "john"]'),
-            status="error",
-            database_name="test_db",
-            schema_name="test_schema",
-            suppression_status="pending",
+            created_at="2022-10-10 10:00:00",
+            updated_at="2022-10-10 10:00:00",
+            status="pending",
+            data=dict(
+                id="1",
+                alert_class_id="elementary.model_id_1",
+                model_unique_id="elementary.model_id_1",
+                alias="modely",
+                path="my/path",
+                original_path="",
+                materialization="table",
+                message="",
+                full_refresh=False,
+                detected_at="2022-10-10 10:00:00",
+                alert_suppression_interval=0,
+                tags='["one", "two"]',
+                model_meta=dict(owner='["jeff", "john"]'),
+                status="error",
+                database_name="test_db",
+                schema_name="test_schema",
+                suppression_status="pending",
+            ),
         ),
-        PendingModelAlertSchema(
+        PendingAlertSchema(
             id="2",
             alert_class_id="elementary.model_id_1",
-            model_unique_id="elementary.model_id_1",
-            alias="modely",
-            path="my/path",
-            original_path="",
-            materialization="table",
-            message="",
-            full_refresh=False,
-            detected_at="2022-10-10 09:00:00",
-            alert_suppression_interval=3,
-            tags='["three"]',
-            model_meta=dict(owner='["john"]'),
-            status="error",
-            database_name="test_db",
-            schema_name="test_schema",
-            suppression_status="pending",
+            type=AlertTypes.MODEL,
+            detected_at="2022-10-10 10:00:00",
+            created_at="2022-10-10 10:00:00",
+            updated_at="2022-10-10 10:00:00",
+            status="pending",
+            data=dict(
+                id="2",
+                alert_class_id="elementary.model_id_1",
+                model_unique_id="elementary.model_id_1",
+                alias="modely",
+                path="my/path",
+                original_path="",
+                materialization="table",
+                message="",
+                full_refresh=False,
+                detected_at="2022-10-10 09:00:00",
+                alert_suppression_interval=3,
+                tags='["three"]',
+                model_meta=dict(owner='["john"]'),
+                status="error",
+                database_name="test_db",
+                schema_name="test_schema",
+                suppression_status="pending",
+            ),
         ),
-        PendingModelAlertSchema(
+        PendingAlertSchema(
             id="3",
             alert_class_id="elementary.model_id_2",
-            model_unique_id="elementary.model_id_2",
-            alias="model2",
-            path="my/path2",
-            original_path="",
-            materialization="table",
-            message="",
-            full_refresh=False,
+            type=AlertTypes.MODEL,
             detected_at="2022-10-10 08:00:00",
-            alert_suppression_interval=1,
-            tags='["three", "four"]',
-            model_meta=dict(owner='["jeff"]'),
-            status="skipped",
-            database_name="test_db",
-            schema_name="test_schema",
-            suppression_status="pending",
+            created_at="2022-10-10 08:00:00",
+            updated_at="2022-10-10 08:00:00",
+            status="pending",
+            data=dict(
+                id="3",
+                alert_class_id="elementary.model_id_2",
+                model_unique_id="elementary.model_id_2",
+                alias="model2",
+                path="my/path2",
+                original_path="",
+                materialization="table",
+                message="",
+                full_refresh=False,
+                detected_at="2022-10-10 08:00:00",
+                alert_suppression_interval=1,
+                tags='["three", "four"]',
+                model_meta=dict(owner='["jeff"]'),
+                status="skipped",
+                database_name="test_db",
+                schema_name="test_schema",
+                suppression_status="pending",
+            ),
         ),
     ]
     source_freshness_alerts = [
-        PendingSourceFreshnessAlertSchema(
+        PendingAlertSchema(
             id="1",
-            source_freshness_execution_id="1",
             alert_class_id="elementary.model_id_1",
-            model_unique_id="elementary.model_id_1",
-            alias="modely",
-            path="my/path",
-            original_path="",
-            materialization="table",
-            message="",
-            full_refresh=False,
-            detected_at="2022-10-10 10:00:00",
-            alert_suppression_interval=0,
-            tags='["one", "two"]',
-            model_meta=dict(owner='["jeff", "john"]'),
-            original_status="error",
-            status="fail",
-            snapshotted_at="2023-08-15T12:26:06.884065+00:00",
-            max_loaded_at="1969-12-31T00:00:00+00:00",
-            max_loaded_at_time_ago_in_s=1692188766.884065,
-            source_name="elementary_integration_tests",
-            identifier="any_type_column_anomalies_validation",
-            error_after='{"count": null, "period": null}',
-            warn_after='{"count": 1, "period": "minute"}',
-            filter="null",
-            error="problemz",
-            database_name="test_db",
-            schema_name="test_schema",
-            suppression_status="pending",
+            type=AlertTypes.SOURCE_FRESHNESS,
+            detected_at="2022-10-10 08:00:00",
+            created_at="2022-10-10 08:00:00",
+            updated_at="2022-10-10 08:00:00",
+            status="pending",
+            data=dict(
+                id="1",
+                source_freshness_execution_id="1",
+                alert_class_id="elementary.model_id_1",
+                model_unique_id="elementary.model_id_1",
+                alias="modely",
+                path="my/path",
+                original_path="",
+                materialization="table",
+                message="",
+                full_refresh=False,
+                detected_at="2022-10-10 10:00:00",
+                alert_suppression_interval=0,
+                tags='["one", "two"]',
+                model_meta=dict(owner='["jeff", "john"]'),
+                original_status="error",
+                status="fail",
+                snapshotted_at="2023-08-15T12:26:06.884065+00:00",
+                max_loaded_at="1969-12-31T00:00:00+00:00",
+                max_loaded_at_time_ago_in_s=1692188766.884065,
+                source_name="elementary_integration_tests",
+                identifier="any_type_column_anomalies_validation",
+                error_after='{"count": null, "period": null}',
+                warn_after='{"count": 1, "period": "minute"}',
+                filter="null",
+                error="problemz",
+                database_name="test_db",
+                schema_name="test_schema",
+                suppression_status="pending",
+            ),
         ),
-        PendingSourceFreshnessAlertSchema(
+        PendingAlertSchema(
             id="2",
-            source_freshness_execution_id="2",
             alert_class_id="elementary.model_id_2",
-            model_unique_id="elementary.model_id_2",
-            alias="modely",
-            path="my/path",
-            original_path="",
-            materialization="table",
-            message="",
-            full_refresh=False,
-            detected_at="2022-10-10 10:00:00",
-            alert_suppression_interval=0,
-            tags='["one", "two"]',
-            model_meta=dict(owner='["jeff", "john"]'),
-            status="warn",
-            original_status="warn",
-            snapshotted_at="2023-08-15T12:26:06.884065+00:00",
-            max_loaded_at="1969-12-31T00:00:00+00:00",
-            max_loaded_at_time_ago_in_s=1692188766.884065,
-            source_name="elementary_integration_tests",
-            identifier="any_type_column_anomalies_validation",
-            error_after='{"count": null, "period": null}',
-            warn_after='{"count": 1, "period": "minute"}',
-            filter="null",
-            error="problemz",
-            database_name="test_db",
-            schema_name="test_schema",
-            suppression_status="pending",
+            type=AlertTypes.SOURCE_FRESHNESS,
+            detected_at="2022-10-10 08:00:00",
+            created_at="2022-10-10 08:00:00",
+            updated_at="2022-10-10 08:00:00",
+            status="pending",
+            data=dict(
+                id="2",
+                source_freshness_execution_id="2",
+                alert_class_id="elementary.model_id_2",
+                model_unique_id="elementary.model_id_2",
+                alias="modely",
+                path="my/path",
+                original_path="",
+                materialization="table",
+                message="",
+                full_refresh=False,
+                detected_at="2022-10-10 10:00:00",
+                alert_suppression_interval=0,
+                tags='["one", "two"]',
+                model_meta=dict(owner='["jeff", "john"]'),
+                status="warn",
+                original_status="warn",
+                snapshotted_at="2023-08-15T12:26:06.884065+00:00",
+                max_loaded_at="1969-12-31T00:00:00+00:00",
+                max_loaded_at_time_ago_in_s=1692188766.884065,
+                source_name="elementary_integration_tests",
+                identifier="any_type_column_anomalies_validation",
+                error_after='{"count": null, "period": null}',
+                warn_after='{"count": 1, "period": "minute"}',
+                filter="null",
+                error="problemz",
+                database_name="test_db",
+                schema_name="test_schema",
+                suppression_status="pending",
+            ),
         ),
-        PendingSourceFreshnessAlertSchema(
+        PendingAlertSchema(
             id="3",
-            source_freshness_execution_id="3",
             alert_class_id="elementary.model_id_3",
-            model_unique_id="elementary.model_id_3",
-            alias="modely",
-            path="my/path",
-            original_path="",
-            materialization="table",
-            message="",
-            full_refresh=False,
-            detected_at="2022-10-10 10:00:00",
-            alert_suppression_interval=0,
-            tags='["one", "two"]',
-            model_meta=dict(owner='["jeff", "john"]'),
-            original_status="runtime error",
-            status="error",
-            snapshotted_at="2023-08-15T12:26:06.884065+00:00",
-            max_loaded_at="1969-12-31T00:00:00+00:00",
-            max_loaded_at_time_ago_in_s=1692188766.884065,
-            source_name="elementary_integration_tests",
-            identifier="any_type_column_anomalies_validation",
-            error_after='{"count": null, "period": null}',
-            warn_after='{"count": 1, "period": "minute"}',
-            filter="null",
-            error="problemz",
-            database_name="test_db",
-            schema_name="test_schema",
-            suppression_status="pending",
+            type=AlertTypes.SOURCE_FRESHNESS,
+            detected_at="2022-10-10 08:00:00",
+            created_at="2022-10-10 08:00:00",
+            updated_at="2022-10-10 08:00:00",
+            status="pending",
+            data=dict(
+                id="3",
+                source_freshness_execution_id="3",
+                alert_class_id="elementary.model_id_3",
+                model_unique_id="elementary.model_id_3",
+                alias="modely",
+                path="my/path",
+                original_path="",
+                materialization="table",
+                message="",
+                full_refresh=False,
+                detected_at="2022-10-10 10:00:00",
+                alert_suppression_interval=0,
+                tags='["one", "two"]',
+                model_meta=dict(owner='["jeff", "john"]'),
+                original_status="runtime error",
+                status="error",
+                snapshotted_at="2023-08-15T12:26:06.884065+00:00",
+                max_loaded_at="1969-12-31T00:00:00+00:00",
+                max_loaded_at_time_ago_in_s=1692188766.884065,
+                source_name="elementary_integration_tests",
+                identifier="any_type_column_anomalies_validation",
+                error_after='{"count": null, "period": null}',
+                warn_after='{"count": 1, "period": "minute"}',
+                filter="null",
+                error="problemz",
+                database_name="test_db",
+                schema_name="test_schema",
+                suppression_status="pending",
+            ),
         ),
     ]
     return test_alerts, model_alerts, source_freshness_alerts
